requirements.txt
tencentcloud-sdk-python==3.0.4
scikit-image==0.17.2
opencv-python==4.2.0.32